Parallelization First Steps
Added a thread count parameter to the base class that all succession extensions use. This will allow developers to parallelize (or not) their succession extensions by specifying more than one thread.

Basic parallelization of AgeCohorts method: if a developer has specified that the succession extension uses more than one thread, then the code attempts to execute the succession extension's AgeCohort's method with multiple threads. The thread safety and synchronization of the method depends on the developer having made the required changes to the method.